llm/requirements.txt
google-adk
python-dotenv
psycopg2-binary
requests
pydantic
